Move tests so that they don't get bundled
