+++ tests/test_bitmap_c_interface.py
