+++ tests/test_pgm_load.py
@@ -20,73 +20,58 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 """
-`adafruit_imageload.tests.test_ppm_load`
+`adafruit_imageload.tests.test_bgm_load`
 ====================================================
 
 * Author(s):  Matt Land
 
 """
 import os
-from io import BytesIO
 from unittest import TestCase
-from .. import pnm
-from ..pnm.ppm_ascii import read_three_colors
+from adafruit_imageload import pnm
 from .displayio_shared_bindings import Bitmap_C_Interface, Palette_C_Interface
 
 
-class TestPpmLoad(TestCase):
-    def test_load_works_p3_ascii(self):
+class TestPgmLoad(TestCase):
+    def test_load_works_p2_ascii(self):
         test_file = os.path.join(
             os.path.dirname(__file__),
             "..",
-            "..",
             "examples",
             "images",
-            "netpbm_p3_rgb_ascii.ppm",
+            "netpbm_p2_ascii.pgm",
         )
         with open(test_file, "rb") as file:
             bitmap, palette = pnm.load(
-                file, b"P3", bitmap=Bitmap_C_Interface, palette=Palette_C_Interface
-            )  # type: Bitmap_C_Interface, Palette_C_Interface
-
-        self.assertTrue(isinstance(palette, Palette_C_Interface))
-        self.assertEqual(6, palette.num_colors)
-        palette.validate()
-        #self.fail(str(palette))
+                file, b"P2", bitmap=Bitmap_C_Interface, palette=Palette_C_Interface
+            )
         self.assertTrue(isinstance(bitmap, Bitmap_C_Interface), bitmap)
         self.assertEqual(6, bitmap.colors)
-        self.assertEqual(16, bitmap.width)
-        self.assertEqual(16, bitmap.height)
+        self.assertEqual(8, bitmap.width)
+        self.assertEqual(8, bitmap.height)
         bitmap.validate()
+        self.assertEqual(6, palette.num_colors)
+        palette.validate()
+        # self.fail(str(palette))
 
-    def test_load_works_p6_binary(self):
+    def test_load_works_p5_binary(self):
         test_file = os.path.join(
             os.path.dirname(__file__),
             "..",
-            "..",
             "examples",
             "images",
-            "netpbm_p6_binary.ppm",
+            "netpbm_p5_binary.pgm",
         )
         with open(test_file, "rb") as file:
             bitmap, palette = pnm.load(
-                file, b"P6", bitmap=Bitmap_C_Interface, palette=Palette_C_Interface
-            )  # type: Bitmap_C_Interface, Palette_C_Interface
-        self.assertEqual(6, palette.num_colors)
-        palette.validate()
+                file, b"P5", bitmap=Bitmap_C_Interface, palette=Palette_C_Interface
+            )
         self.assertTrue(isinstance(bitmap, Bitmap_C_Interface), bitmap)
-        self.assertEqual(6, bitmap.colors)
-        self.assertEqual(16, bitmap.width)
-        self.assertEqual(16, bitmap.height)
+
+        self.assertEqual(8, palette.num_colors)
+        palette.validate()
+        self.assertEqual(8, bitmap.colors)
+        self.assertEqual(8, bitmap.width)
+        self.assertEqual(8, bitmap.height)
         bitmap.validate()
-
-    def test_load_three_colors_tail(self):
-        buffer = BytesIO(b'211 222 233')
-        for i in read_three_colors(buffer):
-            self.assertEqual(b'\xd3\xde\xe9', i)
-
-    def test_load_three_colors_middle(self):
-        buffer = BytesIO(b'0 128 255 45 55 25')
-        for i in iter(read_three_colors(buffer)):
-            self.assertEqual(b'\x00\x80\xff', i)
-            break
+        # self.fail(str(bitmap))
