+++ tests/test_ppm_load.py
@@ -20,60 +20,71 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 """
-`adafruit_imageload.tests.test_bgm_load`
+`adafruit_imageload.tests.test_ppm_load`
 ====================================================
 
 * Author(s):  Matt Land
 
 """
 import os
+from io import BytesIO
 from unittest import TestCase
-from .. import pnm
+from adafruit_imageload import pnm
+from adafruit_imageload.pnm.ppm_ascii import read_three_colors
 from .displayio_shared_bindings import Bitmap_C_Interface, Palette_C_Interface
 
 
-class TestPgmLoad(TestCase):
-    def test_load_works_p2_ascii(self):
+class TestPpmLoad(TestCase):
+    def test_load_works_p3_ascii(self):
         test_file = os.path.join(
             os.path.dirname(__file__),
             "..",
-            "..",
             "examples",
             "images",
-            "netpbm_p2_ascii.pgm",
+            "netpbm_p3_rgb_ascii.ppm",
         )
         with open(test_file, "rb") as file:
             bitmap, palette = pnm.load(
-                file, b"P2", bitmap=Bitmap_C_Interface, palette=Palette_C_Interface
-            )
-        self.assertTrue(isinstance(bitmap, Bitmap_C_Interface), bitmap)
-        self.assertEqual(6, bitmap.colors)
-        self.assertEqual(8, bitmap.width)
-        self.assertEqual(8, bitmap.height)
-        bitmap.validate()
+                file, b"P3", bitmap=Bitmap_C_Interface, palette=Palette_C_Interface
+            )  # type: Bitmap_C_Interface, Palette_C_Interface
+
+        self.assertTrue(isinstance(palette, Palette_C_Interface))
         self.assertEqual(6, palette.num_colors)
         palette.validate()
-        # self.fail(str(palette))
+        #self.fail(str(palette))
+        self.assertTrue(isinstance(bitmap, Bitmap_C_Interface), bitmap)
+        self.assertEqual(6, bitmap.colors)
+        self.assertEqual(16, bitmap.width)
+        self.assertEqual(16, bitmap.height)
+        bitmap.validate()
 
-    def test_load_works_p5_binary(self):
+    def test_load_works_p6_binary(self):
         test_file = os.path.join(
             os.path.dirname(__file__),
             "..",
-            "..",
             "examples",
             "images",
-            "netpbm_p5_binary.pgm",
+            "netpbm_p6_binary.ppm",
         )
         with open(test_file, "rb") as file:
             bitmap, palette = pnm.load(
-                file, b"P5", bitmap=Bitmap_C_Interface, palette=Palette_C_Interface
-            )
-        self.assertTrue(isinstance(bitmap, Bitmap_C_Interface), bitmap)
-
-        self.assertEqual(8, palette.num_colors)
+                file, b"P6", bitmap=Bitmap_C_Interface, palette=Palette_C_Interface
+            )  # type: Bitmap_C_Interface, Palette_C_Interface
+        self.assertEqual(6, palette.num_colors)
         palette.validate()
-        self.assertEqual(8, bitmap.colors)
-        self.assertEqual(8, bitmap.width)
-        self.assertEqual(8, bitmap.height)
+        self.assertTrue(isinstance(bitmap, Bitmap_C_Interface), bitmap)
+        self.assertEqual(6, bitmap.colors)
+        self.assertEqual(16, bitmap.width)
+        self.assertEqual(16, bitmap.height)
         bitmap.validate()
-        # self.fail(str(bitmap))
+
+    def test_load_three_colors_tail(self):
+        buffer = BytesIO(b'211 222 233')
+        for i in read_three_colors(buffer):
+            self.assertEqual(b'\xd3\xde\xe9', i)
+
+    def test_load_three_colors_middle(self):
+        buffer = BytesIO(b'0 128 255 45 55 25')
+        for i in iter(read_three_colors(buffer)):
+            self.assertEqual(b'\x00\x80\xff', i)
+            break
